feat: Change PIDs temperature units for 2.0 GME

+++ src/test/java/org/obd/metrics/codec/giulia_2_0_gme/AthmosphericPressureTest.java
@@ -10,6 +10,6 @@ public class AthmosphericPressureTest implements Giulia_2_0_GME_Test {
 			"62195603F1=1009",
 			}, delimiter = '=')
 	public void parameterizedTest(String input, String expected) {
-		assertEquals(Boolean.TRUE, input, Double.parseDouble(expected));
+		assertEquals(Boolean.TRUE, input, Integer.parseInt(expected));
 	}
 }
